Split utils from helper functions
